tests/wpt/web-platform-tests/IndexedDB/interfaces.htm

<!--
Test converted from WebKit:
http://trac.webkit.org/browser/trunk/LayoutTests/storage/indexeddb/resources/interfaces.js
-->

<!DOCTYPE html>
<!-- Submitted from TestTWF Paris -->
<meta charset=utf-8>
<title>Test IndexedDB's interfaces</title>
<link rel=author href="mailto:romain.huet@gmail.com" title="Romain Huet">

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=support.js></script>

<script>

    test(function() {
        assert_true('IDBCursor' in self);
    }, "Test IDBCursor");

    test(function() {
        assert_true('IDBCursorWithValue' in self);
    }, "Test IDBCursorWithValue");

    test(function() {
        assert_true('IDBDatabase' in self);
    }, "Test IDBDatabase");

    test(function() {
        assert_true('IDBFactory' in self);
    }, "Test IDBFactory");

    test(function() {
        assert_true('IDBIndex' in self);
    }, "Test IDBIndex");

    test(function() {
        assert_true('IDBKeyRange' in self);
    }, "Test IDBKeyRange");

    test(function() {
        assert_true('IDBObjectStore' in self);
    }, "Test IDBObjectStore");

    test(function() {
        assert_true('IDBOpenDBRequest' in self);
    }, "Test IDBOpenDBRequest");

    test(function() {
        assert_true('IDBRequest' in self);
    }, "Test IDBRequest");

    test(function() {
        assert_true('IDBTransaction' in self);
    }, "Test IDBTransaction");

    test(function() {
        assert_true('IDBVersionChangeEvent' in self);
    }, "Test IDBVersionChangeEvent");

</script>

<div id=log></div>
